Fix member IDs inconsistent across runs.
Fixes using '__hash__()' to generate member IDs.

As pointed out in the feedback to the final submission of this
project, the Member ID implementation was flawed in that it
returned different IDs for each program run.

This was due to using 'Member.__hash__()' to generate the IDs.
The method uses a different randomized seed for each execution.

Quoted from the Python 3 documentation;

    Hash randomisation is turned on by default in Python 3.
    This is a security feature:
    Hash randomization is intended to provide protection against a
    denial-of-service caused by carefully-chosen inputs that exploit
    the worst case performance of a dict construction

This method should not be used for IDs that should be consistent
across program execution runs, serialization, etc.

This commit fixes this by instead calculating a MD5 sum of the same
three attributes.